complete renaming with css update
