+++ src/atoms/formTextarea/formTextarea.css
@@ -1,4 +1,4 @@
-/*  title:  textarea
+/*  title:  Form-textarea
     notes:  atom styles
  ==================================================================================================================== */
 
@@ -8,35 +8,35 @@
 /*  variables
 ===================================================================================================================== */
 :root {
-    --Textarea-border: 1px solid rgb(var(--Global-border-color, 170 176 195));
-    --Textarea-border-radius: var(--Global-border-radius, 5px);
-    --Textarea-font-size: 16px;
-    --Textarea-color-error: var(--Global-color-error, 255 0 0);
-    --Textarea-color-success: var(--Global-color-success, 0 255 0);
-    --Textarea-padding: 16px 8px;
+    --Form-textarea-border: 1px solid rgb(var(--Global-border-color, 170 176 195));
+    --Form-textarea-border-radius: var(--Global-border-radius, 5px);
+    --Form-textarea-font-size: 16px;
+    --Form-textarea-color-error: var(--Global-color-error, 255 0 0);
+    --Form-textarea-color-success: var(--Global-color-success, 0 255 0);
+    --Form-textarea-padding: 16px 8px;
 }
 
 
 /* base
 ===================================================================================================================== */
-.Textarea {
-    border: var(--Textarea-border);
-    border-radius: var(--Textarea-border-radius);
-    font-size: var(--Textarea-font-size);
+.Form-textarea {
+    border: var(--Form-textarea-border);
+    border-radius: var(--Form-textarea-border-radius);
+    font-size: var(--Form-textarea-font-size);
     line-height: 1;
     min-height: 80px;
     outline: none;      /* poor accessability */
-    padding: var(--Textarea-padding);
+    padding: var(--Form-textarea-padding);
     transition: .3s;
 }
 
     /* states
     ================================================================================================================= */
-    .Textarea:hover,
-    .Textarea:focus,
-    .Textarea:focus-visible,
-    .Textarea:focus-within,
-    .Textarea:active {
+    .Form-textarea:hover,
+    .Form-textarea:focus,
+    .Form-textarea:focus-visible,
+    .Form-textarea:focus-within,
+    .Form-textarea:active {
         box-shadow: inset 0 0 0 2px #424E79;
         border-color: transparent;
         transition: .2s;
@@ -44,19 +44,19 @@
 
     /* error
     ================================================================================================================= */
-    .Textarea.error {
-        border-color: rgb(var(--Textarea-color-error));
+    .Form-textarea.error {
+        border-color: rgb(var(--Form-textarea-color-error));
     }
 
     /* success
     ================================================================================================================= */
-    .Textarea.success {
-        border-color: rgb(var(--Textarea-color-success));
+    .Form-textarea.success {
+        border-color: rgb(var(--Form-textarea-color-success));
     }
 
     /* disabled
     ================================================================================================================= */
-    .Textarea:disabled {
+    .Form-textarea:disabled {
         background-color: #EFEFEF;
         border: 1px solid #AAB0C3;
         box-shadow: none;
@@ -65,6 +65,6 @@
 
     /* large
     ================================================================================================================= */
-    .Textarea.large {
+    .Form-textarea.large {
         font-size: 22px;
     }
